+++ CHANGELOG.md
@@ -11,12 +11,14 @@
 ### Bug fixes
 
 - [#21][GL-21] - Team invite menu does not validate user input properly
+- [#23][GL-23] - Non-tool items break when mining in the Randomized Drops and Vein Miner scenarios
 - [#29][GL-29] - Teams are sometimes not correctly updated on the tab list
 - [#45][GL-45] - Errors are spammed to the server console on Minecraft 1.19.3+ servers if ProtocolLib is installed
 
 [GL-18]: https://gitlab.com/uhccore/uhccore/-/issues/18
 [GL-20]: https://gitlab.com/uhccore/uhccore/-/issues/20
 [GL-21]: https://gitlab.com/uhccore/uhccore/-/issues/21
+[GL-23]: https://gitlab.com/uhccore/uhccore/-/issues/23
 [GL-29]: https://gitlab.com/uhccore/uhccore/-/issues/29
 [GL-45]: https://gitlab.com/uhccore/uhccore/-/issues/45